refactor(WinningCondition):WinningCondition의 디렉터리를 model로 이동
