+++ src/test/java/lotto/model/WinningRepositoryTest.java
@@ -2,7 +2,6 @@ package lotto.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import lotto.config.WinningCondition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
